refactor: update 메서드 수정 및 컨트롤러 이름 변경
