+++ backend/src/main/java/com/woowacourse/moamoa/studyroom/domain/Article.java
@@ -57,7 +57,5 @@ public abstract class Article extends BaseEntity {
         return studyRoom;
     }
 
-    public abstract void update(Accessor accessor, String title, String content);
-
     public abstract void update(Accessor accessor, Content<?> content);
 }
