+++ backend/src/test/java/com/woowacourse/moamoa/studyroom/controller/DeletingPostArticleControllerTest.java
@@ -1,5 +1,6 @@
 package com.woowacourse.moamoa.studyroom.controller;
 
+import static com.woowacourse.moamoa.studyroom.domain.ArticleType.COMMUNITY;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -12,30 +13,25 @@ import com.woowacourse.moamoa.study.domain.repository.StudyRepository;
 import com.woowacourse.moamoa.study.service.StudyService;
 import com.woowacourse.moamoa.study.service.request.CreatingStudyRequestBuilder;
 import com.woowacourse.moamoa.studyroom.domain.Article;
-import com.woowacourse.moamoa.studyroom.domain.ArticleType;
-import com.woowacourse.moamoa.studyroom.domain.postarticle.CommunityArticle;
-import com.woowacourse.moamoa.studyroom.domain.StudyRoom;
 import com.woowacourse.moamoa.studyroom.domain.repository.studyroom.StudyRoomRepository;
 import com.woowacourse.moamoa.studyroom.domain.repository.article.ArticleRepository;
 import com.woowacourse.moamoa.studyroom.domain.repository.article.ArticleRepositoryFactory;
+import com.woowacourse.moamoa.studyroom.query.LinkArticleDao;
 import com.woowacourse.moamoa.studyroom.query.PostArticleDao;
 import com.woowacourse.moamoa.studyroom.service.ArticleService;
 import com.woowacourse.moamoa.studyroom.service.exception.ArticleNotFoundException;
 import com.woowacourse.moamoa.studyroom.service.exception.UneditableArticleException;
 import com.woowacourse.moamoa.studyroom.service.request.PostArticleRequest;
 import java.time.LocalDate;
-import java.util.Set;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 
 @RepositoryTest
-public class UpdatingArticleControllerTest {
+public class DeletingPostArticleControllerTest {
 
-    CreatingStudyRequestBuilder javaStudyBuilder = new CreatingStudyRequestBuilder()
+    CreatingStudyRequestBuilder javaStudyRequest = new CreatingStudyRequestBuilder()
             .title("java 스터디").excerpt("자바 설명").thumbnail("java image").description("자바 소개");
 
     @Autowired
@@ -53,80 +49,69 @@ public class UpdatingArticleControllerTest {
     @Autowired
     private PostArticleDao postArticleDao;
 
+    @Autowired
+    private LinkArticleDao linkArticleDao;
+
     private StudyService studyService;
-    private ArticleController sut;
+    private PostArticleController sut;
     private ArticleService articleService;
 
     @BeforeEach
     void setUp() {
         studyService = new StudyService(studyRepository, memberRepository, new DateTimeSystem());
         articleService = new ArticleService(studyRoomRepository,
-                articleRepositoryFactory, postArticleDao);
-        sut = new ArticleController(articleService);
+                articleRepositoryFactory, postArticleDao, linkArticleDao);
+        sut = new PostArticleController(articleService);
     }
 
-    @DisplayName("게시글을 수정한다.")
+    @DisplayName("스터디 커뮤니티 게시글을 삭제한다.")
     @Test
-    void updateArticle() {
+    void deleteCommunityArticle() {
         // arrange
-        Member owner = memberRepository.save(new Member(1L, "username", "image", "profile"));
+        Member member = memberRepository.save(new Member(1L, "username", "imageUrl", "profileUrl"));
+
         Study study = studyService
-                .createStudy(owner.getGithubId(), javaStudyBuilder.startDate(LocalDate.now()).build());
-        Article article = articleService
-                .createArticle(owner.getId(), study.getId(), new PostArticleRequest("제목", "내용"), ArticleType.COMMUNITY);
+                .createStudy(member.getGithubId(), javaStudyRequest.startDate(LocalDate.now()).build());
 
-        // act
-        final ResponseEntity<Void> response = sut
-                .updateArticle(owner.getId(), study.getId(), article.getId(), ArticleType.COMMUNITY,
-                        new PostArticleRequest("제목 수정", "내용 수정"));
+        PostArticleRequest request = new PostArticleRequest("게시글 제목", "게시글 내용");
+        Article article = articleService.createArticle(member.getId(), study.getId(), request, COMMUNITY);
 
-        // assert
-        ArticleRepository<Article> articleRepository = articleRepositoryFactory.getRepository(ArticleType.COMMUNITY);
-        Article actualArticle = articleRepository.findById(article.getId()).orElseThrow();
+        //act
+        sut.deleteArticle(member.getId(), study.getId(), article.getId(), COMMUNITY);
 
-        StudyRoom expectStudyRoom = new StudyRoom(study.getId(), owner.getId(), Set.of());
-        CommunityArticle expectArticle = new CommunityArticle(article.getId(), "제목 수정", "내용 수정", owner.getId(),
-                expectStudyRoom);
-
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.NO_CONTENT);
-        assertThat(actualArticle).isEqualTo(expectArticle);
+        //assert
+        ArticleRepository<Article> articleRepository = articleRepositoryFactory.getRepository(COMMUNITY);
+        assertThat(articleRepository.existsById(article.getId())).isFalse();
     }
 
-    @DisplayName("게시글이 없는 경우 수정 시 예외가 발생한다.")
+    @DisplayName("게시글이 없는 경우 조회 시 예외가 발생한다.")
     @Test
-    void throwExceptionWhenUpdateToNotFoundArticle() {
+    void throwExceptionWhenGettingToNotFoundArticle() {
         // arrange
         Member member = memberRepository.save(new Member(1L, "username", "imageUrl", "profileUrl"));
         Study study = studyService
-                .createStudy(member.getGithubId(), javaStudyBuilder.startDate(LocalDate.now()).build());
+                .createStudy(member.getGithubId(), javaStudyRequest.startDate(LocalDate.now()).build());
 
         // act & assert
-        assertThatThrownBy(() ->
-                sut.updateArticle(member.getId(), study.getId(), 1L, ArticleType.COMMUNITY,
-                        new PostArticleRequest("제목 수정", "내용 수정"))
-        )
+        assertThatThrownBy(() -> sut.deleteArticle(member.getId(), study.getId(), 1L, COMMUNITY))
                 .isInstanceOf(ArticleNotFoundException.class);
     }
 
-    @DisplayName("게시글을 수정할 수 없는 경우 예외가 발생한다.")
+    @DisplayName("게시글을 삭제할 수 없는 경우 예외가 발생한다.")
     @Test
-    void throwExceptionWhenUpdateByNotParticipant() {
+    void throwExceptionWhenDeletingByNotParticipant() {
         // arrange
         Member member = memberRepository.save(new Member(1L, "username", "imageUrl", "profileUrl"));
         Member other = memberRepository.save(new Member(2L, "username2", "imageUrl", "profileUrl"));
 
         Study study = studyService
-                .createStudy(member.getGithubId(), javaStudyBuilder.startDate(LocalDate.now()).build());
+                .createStudy(member.getGithubId(), javaStudyRequest.startDate(LocalDate.now()).build());
 
         PostArticleRequest request = new PostArticleRequest("게시글 제목", "게시글 내용");
-        final Article article = articleService
-                .createArticle(member.getId(), study.getId(), request, ArticleType.COMMUNITY);
+        Article article = articleService.createArticle(member.getId(), study.getId(), request, COMMUNITY);
 
         // act & assert
-        assertThatThrownBy(() ->
-                sut.updateArticle(other.getId(), study.getId(), article.getId(), ArticleType.COMMUNITY,
-                        new PostArticleRequest("제목 수정", "내용 수정"))
-        )
+        assertThatThrownBy(() -> sut.deleteArticle(other.getId(), study.getId(), article.getId(), COMMUNITY))
                 .isInstanceOf(UneditableArticleException.class);
     }
 }
