+++ backend/src/test/java/com/woowacourse/moamoa/studyroom/controller/UpdatingPostArticleControllerTest.java
@@ -1,6 +1,5 @@
 package com.woowacourse.moamoa.studyroom.controller;
 
-import static com.woowacourse.moamoa.studyroom.domain.ArticleType.COMMUNITY;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -13,24 +12,31 @@ import com.woowacourse.moamoa.study.domain.repository.StudyRepository;
 import com.woowacourse.moamoa.study.service.StudyService;
 import com.woowacourse.moamoa.study.service.request.CreatingStudyRequestBuilder;
 import com.woowacourse.moamoa.studyroom.domain.Article;
+import com.woowacourse.moamoa.studyroom.domain.ArticleType;
+import com.woowacourse.moamoa.studyroom.domain.postarticle.CommunityArticle;
+import com.woowacourse.moamoa.studyroom.domain.StudyRoom;
 import com.woowacourse.moamoa.studyroom.domain.repository.studyroom.StudyRoomRepository;
 import com.woowacourse.moamoa.studyroom.domain.repository.article.ArticleRepository;
 import com.woowacourse.moamoa.studyroom.domain.repository.article.ArticleRepositoryFactory;
+import com.woowacourse.moamoa.studyroom.query.LinkArticleDao;
 import com.woowacourse.moamoa.studyroom.query.PostArticleDao;
 import com.woowacourse.moamoa.studyroom.service.ArticleService;
 import com.woowacourse.moamoa.studyroom.service.exception.ArticleNotFoundException;
 import com.woowacourse.moamoa.studyroom.service.exception.UneditableArticleException;
 import com.woowacourse.moamoa.studyroom.service.request.PostArticleRequest;
 import java.time.LocalDate;
+import java.util.Set;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 
 @RepositoryTest
-public class DeletingArticleControllerTest {
+public class UpdatingPostArticleControllerTest {
 
-    CreatingStudyRequestBuilder javaStudyRequest = new CreatingStudyRequestBuilder()
+    CreatingStudyRequestBuilder javaStudyBuilder = new CreatingStudyRequestBuilder()
             .title("java 스터디").excerpt("자바 설명").thumbnail("java image").description("자바 소개");
 
     @Autowired
@@ -48,66 +54,83 @@ public class DeletingArticleControllerTest {
     @Autowired
     private PostArticleDao postArticleDao;
 
+    @Autowired
+    private LinkArticleDao linkArticleDao;
+
     private StudyService studyService;
-    private ArticleController sut;
+    private PostArticleController sut;
     private ArticleService articleService;
 
     @BeforeEach
     void setUp() {
         studyService = new StudyService(studyRepository, memberRepository, new DateTimeSystem());
         articleService = new ArticleService(studyRoomRepository,
-                articleRepositoryFactory, postArticleDao);
-        sut = new ArticleController(articleService);
+                articleRepositoryFactory, postArticleDao, linkArticleDao);
+        sut = new PostArticleController(articleService);
     }
 
-    @DisplayName("스터디 커뮤니티 게시글을 삭제한다.")
+    @DisplayName("게시글을 수정한다.")
     @Test
-    void deleteCommunityArticle() {
+    void updateArticle() {
         // arrange
-        Member member = memberRepository.save(new Member(1L, "username", "imageUrl", "profileUrl"));
-
+        Member owner = memberRepository.save(new Member(1L, "username", "image", "profile"));
         Study study = studyService
-                .createStudy(member.getGithubId(), javaStudyRequest.startDate(LocalDate.now()).build());
+                .createStudy(owner.getGithubId(), javaStudyBuilder.startDate(LocalDate.now()).build());
+        Article article = articleService
+                .createArticle(owner.getId(), study.getId(), new PostArticleRequest("제목", "내용"), ArticleType.COMMUNITY);
 
-        PostArticleRequest request = new PostArticleRequest("게시글 제목", "게시글 내용");
-        Article article = articleService.createArticle(member.getId(), study.getId(), request, COMMUNITY);
+        // act
+        final ResponseEntity<Void> response = sut
+                .updateArticle(owner.getId(), study.getId(), article.getId(), ArticleType.COMMUNITY,
+                        new PostArticleRequest("제목 수정", "내용 수정"));
 
-        //act
-        sut.deleteArticle(member.getId(), study.getId(), article.getId(), COMMUNITY);
+        // assert
+        ArticleRepository<Article> articleRepository = articleRepositoryFactory.getRepository(ArticleType.COMMUNITY);
+        Article actualArticle = articleRepository.findById(article.getId()).orElseThrow();
 
-        //assert
-        ArticleRepository<Article> articleRepository = articleRepositoryFactory.getRepository(COMMUNITY);
-        assertThat(articleRepository.existsById(article.getId())).isFalse();
+        StudyRoom expectStudyRoom = new StudyRoom(study.getId(), owner.getId(), Set.of());
+        CommunityArticle expectArticle = new CommunityArticle(article.getId(), "제목 수정", "내용 수정", owner.getId(),
+                expectStudyRoom);
+
+        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.NO_CONTENT);
+        assertThat(actualArticle).isEqualTo(expectArticle);
     }
 
-    @DisplayName("게시글이 없는 경우 조회 시 예외가 발생한다.")
+    @DisplayName("게시글이 없는 경우 수정 시 예외가 발생한다.")
     @Test
-    void throwExceptionWhenGettingToNotFoundArticle() {
+    void throwExceptionWhenUpdateToNotFoundArticle() {
         // arrange
         Member member = memberRepository.save(new Member(1L, "username", "imageUrl", "profileUrl"));
         Study study = studyService
-                .createStudy(member.getGithubId(), javaStudyRequest.startDate(LocalDate.now()).build());
+                .createStudy(member.getGithubId(), javaStudyBuilder.startDate(LocalDate.now()).build());
 
         // act & assert
-        assertThatThrownBy(() -> sut.deleteArticle(member.getId(), study.getId(), 1L, COMMUNITY))
+        assertThatThrownBy(() ->
+                sut.updateArticle(member.getId(), study.getId(), 1L, ArticleType.COMMUNITY,
+                        new PostArticleRequest("제목 수정", "내용 수정"))
+        )
                 .isInstanceOf(ArticleNotFoundException.class);
     }
 
-    @DisplayName("게시글을 삭제할 수 없는 경우 예외가 발생한다.")
+    @DisplayName("게시글을 수정할 수 없는 경우 예외가 발생한다.")
     @Test
-    void throwExceptionWhenDeletingByNotParticipant() {
+    void throwExceptionWhenUpdateByNotParticipant() {
         // arrange
         Member member = memberRepository.save(new Member(1L, "username", "imageUrl", "profileUrl"));
         Member other = memberRepository.save(new Member(2L, "username2", "imageUrl", "profileUrl"));
 
         Study study = studyService
-                .createStudy(member.getGithubId(), javaStudyRequest.startDate(LocalDate.now()).build());
+                .createStudy(member.getGithubId(), javaStudyBuilder.startDate(LocalDate.now()).build());
 
         PostArticleRequest request = new PostArticleRequest("게시글 제목", "게시글 내용");
-        Article article = articleService.createArticle(member.getId(), study.getId(), request, COMMUNITY);
+        final Article article = articleService
+                .createArticle(member.getId(), study.getId(), request, ArticleType.COMMUNITY);
 
         // act & assert
-        assertThatThrownBy(() -> sut.deleteArticle(other.getId(), study.getId(), article.getId(), COMMUNITY))
+        assertThatThrownBy(() ->
+                sut.updateArticle(other.getId(), study.getId(), article.getId(), ArticleType.COMMUNITY,
+                        new PostArticleRequest("제목 수정", "내용 수정"))
+        )
                 .isInstanceOf(UneditableArticleException.class);
     }
 }
