+++ backend/src/test/java/com/woowacourse/moamoa/studyroom/webmvc/GettingPostArticleControllerWebMvcTest.java
@@ -13,7 +13,7 @@ import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.http.HttpHeaders;
 
-public class GettingArticleControllerWebMvcTest extends WebMVCTest {
+public class GettingPostArticleControllerWebMvcTest extends WebMVCTest {
 
     @DisplayName("잘못된 토큰으로 커뮤니티 글을 조회할 경우 401을 반환한다.")
     @ParameterizedTest
