+++ backend/src/test/java/com/woowacourse/moamoa/studyroom/webmvc/UpdatingPostArticleControllerWebMvcTest.java
@@ -1,48 +1,42 @@
 package com.woowacourse.moamoa.studyroom.webmvc;
 
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.when;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.woowacourse.moamoa.WebMVCTest;
-import com.woowacourse.moamoa.studyroom.service.ArticleService;
 import com.woowacourse.moamoa.studyroom.service.request.PostArticleRequest;
-import com.woowacourse.moamoa.member.service.exception.NotParticipatedMemberException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 
-public class CreatingArticleControllerWebMvcTest extends WebMVCTest {
+public class UpdatingPostArticleControllerWebMvcTest extends WebMVCTest {
 
-    @MockBean
-    private ArticleService articleService;
-
-    @DisplayName("잘못된 토큰으로 커뮤니티 글을 생성할 경우 401을 반환한다.")
+    @DisplayName("잘못된 토큰으로 커뮤니티 글을 수정할 경우 401을 반환한다.")
     @ParameterizedTest
     @ValueSource(strings = {"", "Bearer InvalidToken", "Invalid"})
     void unauthorizedByInvalidToken(String token) throws Exception {
         mockMvc.perform(
-                post("/api/studies/{study-id}/community/articles", 1L)
+                put("/api/studies/{study-id}/community/articles/{article-id}", 1L, 1L)
                         .header(HttpHeaders.AUTHORIZATION, token)
         )
                 .andExpect(status().isUnauthorized())
                 .andDo(print());
     }
 
-    @DisplayName("스터디 ID가 잘못된 형식인 경우 400에러를 반환한다.")
-    @Test
-    void badRequestByInvalidStudyIdFormat() throws Exception {
+    @DisplayName("스터디 ID 또는 게시글 ID가 잘못된 형식인 경우 400에러를 반환한다.")
+    @ParameterizedTest
+    @CsvSource({"one, 1", "1, one"})
+    void badRequestByInvalidIdFormat(String studyId, String articleId) throws Exception {
         final String token = "Bearer" + tokenProvider.createToken(1L).getAccessToken();
 
         mockMvc.perform(
-                post("/api/studies/{study-id}/community/articles", "one")
+                put("/api/studies/{study-id}/community/articles/{article-id}", studyId, articleId)
                         .header(HttpHeaders.AUTHORIZATION, token)
         )
                 .andExpect(status().isBadRequest())
@@ -56,10 +50,10 @@ public class CreatingArticleControllerWebMvcTest extends WebMVCTest {
         final String token = "Bearer" + tokenProvider.createToken(1L).getAccessToken();
 
         mockMvc.perform(
-                post("/api/studies/{study-id}/community/articles", "1")
+                put("/api/studies/{study-id}/community/articles/{article-id}", "1", "1")
                         .header(HttpHeaders.AUTHORIZATION, token)
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(objectMapper.writeValueAsString(new PostArticleRequest(title, "content")))
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(new PostArticleRequest(title, "content")))
         )
                 .andExpect(status().isBadRequest())
                 .andDo(print());
@@ -72,7 +66,7 @@ public class CreatingArticleControllerWebMvcTest extends WebMVCTest {
         final String token = "Bearer" + tokenProvider.createToken(1L).getAccessToken();
 
         mockMvc.perform(
-                post("/api/studies/{study-id}/community/articles", "1")
+                put("/api/studies/{study-id}/community/articles/{article-id}", "1", "1")
                         .header(HttpHeaders.AUTHORIZATION, token)
                         .contentType(MediaType.APPLICATION_JSON)
                         .content(objectMapper.writeValueAsString(new PostArticleRequest("title", content)))
@@ -88,7 +82,7 @@ public class CreatingArticleControllerWebMvcTest extends WebMVCTest {
         final String token = "Bearer" + tokenProvider.createToken(1L).getAccessToken();
 
         mockMvc.perform(
-                post("/api/studies/{study-id}/community/articles", "1")
+                put("/api/studies/{study-id}/community/articles/{article-id}", "1", "1")
                         .header(HttpHeaders.AUTHORIZATION, token)
                         .contentType(MediaType.APPLICATION_JSON)
                         .content(objectMapper.writeValueAsString(new PostArticleRequest(title, "content")))
@@ -104,7 +98,7 @@ public class CreatingArticleControllerWebMvcTest extends WebMVCTest {
         final String token = "Bearer" + tokenProvider.createToken(1L).getAccessToken();
 
         mockMvc.perform(
-                post("/api/studies/{study-id}/community/articles", "1")
+                put("/api/studies/{study-id}/community/articles/{article-id}", "1", "1")
                         .header(HttpHeaders.AUTHORIZATION, token)
                         .contentType(MediaType.APPLICATION_JSON)
                         .content(objectMapper.writeValueAsString(new PostArticleRequest("title", content)))
@@ -119,11 +113,11 @@ public class CreatingArticleControllerWebMvcTest extends WebMVCTest {
         final String token = "Bearer" + tokenProvider.createToken(1L).getAccessToken();
 
         mockMvc.perform(
-                        post("/api/studies/{study-id}/community/articles", "1")
-                                .header(HttpHeaders.AUTHORIZATION, token)
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(objectMapper.writeValueAsString(new PostArticleRequest("a".repeat(31), "cotent")))
-                )
+                put("/api/studies/{study-id}/community/articles/{article-id}", "1", "1")
+                        .header(HttpHeaders.AUTHORIZATION, token)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(new PostArticleRequest("a".repeat(31), "cotent")))
+        )
                 .andExpect(status().isBadRequest())
                 .andDo(print());
     }
@@ -134,29 +128,12 @@ public class CreatingArticleControllerWebMvcTest extends WebMVCTest {
         final String token = "Bearer" + tokenProvider.createToken(1L).getAccessToken();
 
         mockMvc.perform(
-                        post("/api/studies/{study-id}/community/articles", "1")
-                                .header(HttpHeaders.AUTHORIZATION, token)
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(objectMapper.writeValueAsString(new PostArticleRequest("a".repeat(5001), "cotent")))
-                )
+                put("/api/studies/{study-id}/community/articles/{article-id}", "1", "1")
+                        .header(HttpHeaders.AUTHORIZATION, token)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(new PostArticleRequest("a".repeat(5001), "cotent")))
+        )
                 .andExpect(status().isBadRequest())
-                .andDo(print());
-    }
-    
-    @DisplayName("스터디에 참여한 참가자가 아닌 경우, NotParticipatedMemberException이 발생하고 401을 반환한다.")
-    @Test
-    void unauthorizedByNotParticipant() throws Exception {
-        when(articleService.createArticle(any(), any(), any(), any())).thenThrow(NotParticipatedMemberException.class);
-
-        final String token = "Bearer" + tokenProvider.createToken(1L).getAccessToken();
-
-        mockMvc.perform(
-                        post("/api/studies/{study-id}/community/articles", "1")
-                                .header(HttpHeaders.AUTHORIZATION, token)
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(objectMapper.writeValueAsString(new PostArticleRequest("title", "content")))
-                )
-                .andExpect(status().isUnauthorized())
                 .andDo(print());
     }
 }
